requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
watchdog==3.0.0
aiofiles==23.2.1
python-multipart==0.0.6
cryptography==41.0.5
pyjwt==2.8.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 